requirements.txt
Django==4.1.5
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
phonenumbers==8.13.5
django-environ==0.9.0
